+++ frontend/src/components/TaskList.jsx
@@ -1,5 +1,36 @@
+import { useEffect, useState } from 'react';
+import axios from 'axios';
+
 export const TaskList = () => {
+  const [tasks, setTasks] = useState([]);
+
+  useEffect(() => {
+    getTasks();
+  }, []);
+
+  const getTasks = () => {
+    axios.get('http://localhost:8000/api/tasks')
+      .then(response => {
+        setTasks(response.data)
+        console.log(response.data);
+      })
+      .catch(e => {
+        console.log(e);
+      });
+  };
+
   return (
-    <>一覧ページ</>
-  )
+    <>
+      <h2 className="text-center m-5">Task List</h2>
+      <div class="card w-50 mx-auto">
+        <ul className="list-group">
+          {tasks.map((task, index) => (
+            <li className={"list-group-item"} key={index}>
+              {task.title}
+            </li>
+          ))}
+        </ul>
+      </div>
+    </>
+  );
 }
